+++ src/ascnhalf-world/SpellAuras.cpp
@@ -5243,7 +5243,7 @@ void Aura::EventPeriodicLeech(uint32 amount)
 			}
 		}
 
-		uint32 siphonbonus = auras.size() * m_caster->m_soulSiphon.amt;
+		uint32 siphonbonus = (uint32)(auras.size() * m_caster->m_soulSiphon.amt);
 		if( siphonbonus > m_caster->m_soulSiphon.max )
 			siphonbonus = m_caster->m_soulSiphon.max;
